added more seed data and renamed PostDisplay and TitleList components
