+++ frontend/src/components/Homepage.js
@@ -1,11 +1,11 @@
 import React, { Component } from 'react';
-import PostList from '../containers/PostList';
+import TitleList from '../containers/TitleList';
 
 /** Homepage: displays list of all blog posts */
 class Homepage extends Component {
   render() {
     return (
-      <PostList />
+      <TitleList />
     );
   }
 }
