+++ frontend/src/components/PostDisplay.css
@@ -1,14 +1,14 @@
-.PostTitle {
+.PostDisplay {
   position: relative;
 }
 
-.PostTitle-right {
+.PostDisplay-right {
   position: absolute;
   right: 0;
   top: 1em;
   text-align: right;
 }
 
-.PostTitle-right i {
+.PostDisplay-right i {
   margin-left: 1em;
 }
